requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==3.0.0
attrs==24.3.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.12.3
bip-utils==2.9.3
bitarray==3.0.0
bitcoin==1.1.39
bitcoinlib==0.7.1
bleach==6.2.0
blockcypher==1.0.93
cbor2==5.6.5
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
ckzg==2.0.1
coincurve==20.0.0
colorama==0.4.6
construct==2.10.68
construct-typing==0.5.6
crcmod==1.7
cryptography==41.0.3
cytoolz==1.0.1
dj-database-url==1.0.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
Django==5.1.4
django-cors-headers==4.6.0
django-ninja==1.3.0
docopt==0.6.2
ecdsa==0.19.0
ECPy==1.2.5
ed25519-blake2b==1.4.1
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==2.1.0
eth-typing==5.0.1
eth-utils==5.1.0
eth_abi==5.1.0
executing==2.1.0
fastjsonschema==2.21.1
frozenlist==1.5.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
hexbytes==1.2.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.5
jsonalias==0.1.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.0
mnemonic==0.21
multidict==6.1.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
numpy==2.2.1
packaging==24.2
pandocfilters==1.5.1
parsimonious==0.10.0
parso==0.8.4
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
pure_eval==0.2.3
py-coingecko-client==1.0.0
py-sr25519-bindings==0.2.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pyunormalize==16.0.0
# pywin32==308
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rlp==4.0.1
rpds-py==0.22.3
six==1.17.0
sniffio==1.3.1
solana==0.36.2
solders==0.23.0
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.5.3
stack-data==0.6.3
tinycss2==1.4.0
toolz==1.0.0
tornado==6.4.2
traitlets==5.14.3
tronpy==0.5.0
types-Deprecated==1.2.15.20241117
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
web3==7.6.1
webencodings==0.5.1
websockets==13.1
whitenoise==6.8.2
wrapt==1.17.0
xrpl-py==4.0.0
yarg==0.1.9
yarl==1.18.3
